% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalizers.R
\name{aggregate_pdfs}
\alias{aggregate_pdfs}
\title{Aggregate conditional pdfs into a marginal pdf}
\usage{
aggregate_pdfs(pdfs, probs)
}
\arguments{
\item{pdfs}{list of pdfs}

\item{probs}{numeric vector (or list) of probabilities the same length as \code{pdf_list}}
}
\value{
marginal pdf (returned as a function)
}
\description{
Aggregate conditional pdfs into a marginal pdf
}
