% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphas_inventory_evaluation.R
\name{alphas_inventory_evaluation}
\alias{alphas_inventory_evaluation}
\title{Inventory Evaluation for different alphas}
\usage{
alphas_inventory_evaluation(
  demand_forecast,
  demand_test,
  alpha = 0.05,
  plot.out = FALSE,
  mode = "normal",
  hist_demand = NA,
  alpha.seq = TRUE
)
}
\arguments{
\item{demand_forecast}{The forecast. This can be a forecast object, a list
of mean and sd of the forecast or a point cloud of the distribution of the
forecast.}

\item{demand_test}{A vector containing the actual demand for the inventory
evaluation. The length of the test demand should be equal to the length of
the demand forecast.}

\item{alpha}{A sequence of target service levels of the inventory simulation.
This is can be a single value, or a sequence. If alpha.seq is set on TRUE, a
sequence is generated (by default).}

\item{plot.out}{Parameter to plot a graphic for every alpha in the sequence
, default set on FALSE.}

\item{mode}{Shape of the distribution, default set on 'normal'.}

\item{hist_demand}{A vector of historical demand of the series, where the demand
forecast model is build on. Default is set on NA.}

\item{alpha.seq}{Should the function generate a default alpha sequence. This
is set by default on TRUE.}
}
\value{
A numeric vector with: used service level parameter alpha, stockout
rate, average inventory, average on-hand inventory, fill-rate, average test
demand, ratio average on-hand inventory on average test demand. A plot of
the simulation can be shown.
}
\description{
Evaluates a forecast for a sequence of alphas via an inventory simulation
}
\details{
This function returns the output of a sequence of inventory simulations
with different target alphas. This output can be used to generate a trade-off
curve between fill rate and average on-hand inventory.
}
\examples{
\dontrun{
series <- rnorm(60,100,1)
test_series <- rnorm(12,100,1)
forecast <- series[(length(series)-12):length(series)]
alphas_inventory_evaluation(demand_forecast=list(forecast,sd=rep(3,12)),
demand_test=test_series, hist_demand=series, plot.out=TRUE)
}
}
\author{
Yves R. Sagaert
}
