% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battery.R
\name{battery}
\alias{battery}
\title{Agrup sequencial emissions for record}
\usage{
battery(grid, varible = NA, verbose = TRUE)
}
\arguments{
\item{grid}{list of ouputs from griding funcion}

\item{verbose}{display additional information}

\item{variable}{name(s) of the pollutant(s)}
}
\description{
create stack of grinded emissions for record in a NetCDF file with multiple dates
}
\examples{
grinded_so2 <- sf::read_sf(paste0(system.file("extdata",package="inventory"),"/grid_so2.gpkg"))
two_times_grinded_so2 <- battery(list(a = grinded_so2, b = grinded_so2))
}
