% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRboot.ld.fr.R
\name{CRboot.ld.fr}
\alias{CRboot.ld.fr}
\title{Estimate Bootstrap ROP or SS from Lead-Time and Demand Data for a Continuous Review Policy
with a Target Fill Rate}
\usage{
CRboot.ld.fr(
  x,
  y,
  qty,
  p2,
  B = 500,
  roptru = TRUE,
  seed = as.numeric(Sys.time())
)
}
\arguments{
\item{x}{a vector of lead-times}

\item{y}{a vector of demands}

\item{qty}{the fixed order quantity}

\item{p2}{the target fill rate}

\item{B}{the number of bootstrap resamples, default is 500}

\item{roptru}{if TRUE returns the ROP else the SS, default is the ROP}

\item{seed}{a random seed, default is the computer time}
}
\value{

}
\description{
The function uses one input vector each of lead time and demand to determine the reorder
point (ROP) for a target P2 proportion of demand fulfilled from available stock i.e.
fill rate. for the continuous review inventory control policy. The full algorithm is
provided in Saldanha (2022). The inputs must include lead time and demand vectors a P2 fill
rate and a fixed order quantity (qty). An optional number of B bootstrap resamples may be
provided, the default is set to 500. It is recommended not to set B<200 resamples. As B
increases (e.g. B>1000) the marginal accuracy gained is trivial and the run time will
increase, especially if used in batch runs to determine ROPs over hunderds of items. As per
Saldanha (2022) the ROP is estimated automatically using the traditional fill rate
estimator when P2>0.9 and the Silver (1970) approach otherwise.
}
\details{
The ROP is the default value returned from the function but if safety stock is desired
then an optional input can be added using "roptru=FALSE" will return the safety stock.

Note that the bootstrap approach utilizes a random resampling of the input lead time and
demand resamples with replacement. The random seed for the random resampling is set to the
current machine's system time. Hence, if replicability of the results are desired e.g., the
ROP and safety stock for a number of items need to be calculated then it is recommended to
use the input "seed=" (see examples).
}
\examples{
a<-rnorm(20,5,1)
b<-rnorm(20,20,5)
# Returns the ROP for fixed order qty=50 and 500 bootstrap resamples with a fill rate of 95\%
CRboot.ld.fr(a,b,50,.95,500)
# Returns the SS for fixed order qty=50 and 500 bootstrap resamples with a fill rate of 95\%
CRboot.ld.fr(a,b,50,.95,500,FALSE) # Returns the ROP and SS from the same random numbers
CRboot.ld.fr(a,b,50,.95,500,TRUE,0)
CRboot.ld.fr(a,b,50,.95,500,FALSE,0)
}
