% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRboot.x.fr.R
\name{CRboot.x.fr}
\alias{CRboot.x.fr}
\title{Estimate Bootstrap ROP or SS from LTD Data for a Continuous Review Policy with a
Target Fill Rate}
\usage{
CRboot.x.fr(x, qty, p2, B = 500, roptru = TRUE, seed = as.numeric(Sys.time()))
}
\arguments{
\item{x}{a vector of lead time demands}

\item{qty}{the fixed order quantity}

\item{p2}{the target fill rate}

\item{B}{the number of bootstrap resamples, default is 500}

\item{roptru}{if TRUE returns the ROP else the SS, default is the ROP}

\item{seed}{a random seed, default is the computer time}
}
\value{
the bootstrap estimate ROP or SS
}
\description{
The function uses an input vector of lead time demands (LTD) to determine the reorder
point (ROP) for a target P2 proportion of demand fulfilled from available stock i.e.
fill rate. for the continuous review inventory control policy. The full algorithm is
provided in Saldanha (2022). The inputs must include a LTD vector a P2 fill rate and a
fixed order quantity (qty). An optional number of B bootstrap resamples may be provided,
the default is set to 500. It is recommended not to set B<200 resamples. As B increases
(e.g. B>1000) the marginal accuracy gained is trivial and the run time will increase,
especially if used in batch runs to determine ROPs over hunderds of items. As per
Saldanha (2022) the ROP is estimated automatically using the traditional fill rate
estimator when P2>0.9 and the Silver (1970) approach otherwise.
}
\details{
The ROP is the default value returned from the function but if safety stock is desired
then an optional input can be added where "roptru=FALSE" will return the safety stock.

Note that the bootstrap approach utilizes a random resampling of the input LTD resamples
with replacement. The random seed for the random resampling is set to the current machine's
system time. Hence, if replicability of the results are desired e.g., the ROP and safety
stock for a number of items need to be calculated then it is recommended to use the input
"seed=" (see examples).
}
\examples{
a<-rnorm(30,100,20)
Returns ROP for fixed order qty=50 and 500 bootstrap resamples with a fill rate of 95\%
CRboot.x.fr(a,50,.95,500)
Returns the SS for fixed order qty=50 and 500 bootstrap resamples with a fill rate of 95\%
CRboot.x.fr(a,50,.95,500,FALSE)
Returns the ROP and SS from the same random numbers
CRboot.x.fr(a,50,.95,500,TRUE,0)
CRboot.x.fr(a,50,.95,500,FALSE,0)
}
