% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRgamma.x.csl.R
\name{CRgamma.x.csl}
\alias{CRgamma.x.csl}
\title{Gamma Cycle Service Level ROP and SS Calculation from lead time demand data under the
Continuous Review Inventory Control Policy}
\usage{
CRgamma.x.csl(x, p1, roptru = TRUE)
}
\arguments{
\item{x}{vector of lead time demands}

\item{p1}{the cycle service level as PNS}

\item{roptru}{if TRUE returns the ROP else SS, default is TRUE}
}
\value{
The gamma approximation ROP or SS
}
\description{
This function uses the input of a vector of lead time demand to calculate
the reorder point (ROP) or safety stock (SS) for a single inventory item with a P1
probability of no stockout (PNS) under the (s, Q) inventory control policy using the gamma
approximation. The gamma approximation assumes that lead time demand is distributed according
to the gamma distribution. This function is available in popular textbooks e.g. Silver, Pyke
and Thomas (2016, 311-312).
}
\details{
The output is the single inventory items' gamma ROP (s) if the roptru parameter is TRUE
or not entered (default) else if FALSE the SS for the target PNS P1 is returned.
}
\examples{
CRgamma.x.csl(c(runif(24,200,400)),0.90)
}
