% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRlnorm.ld.fr.R, R/CRlnorm.x.fr.R
\name{CRlnorm.ld.fr}
\alias{CRlnorm.ld.fr}
\title{Lognormal Fill Rate ROP and SS Calculation from lead-time and demand data for the
Continuous Review Inventory Control Policy}
\usage{
CRlnorm.ld.fr(x, p2, qty, roptru = TRUE)

CRlnorm.ld.fr(x, p2, qty, roptru = TRUE)
}
\arguments{
\item{x}{vector of lead time demands}

\item{p2}{item's fill rate target proportion demand fulfilled from available stock}

\item{qty}{item's fixed order quantity}

\item{roptru}{if TRUE returns the ROP else SS, default is TRUE}

\item{y}{time series of demands in the same unit of time as lead time}
}
\value{
The lognormal approximation ROP or SS

the lognormal approximation reorder point (ROP)
}
\description{
This function uses the input of a vector of lead time and a vector of demand to calculate
the reorder point (ROP) or safety stock (SS) for a single inventory item with a P2
proportion of demand fulfilled from available stock fill rate under the (s, Q) inventory
control policy using the lognormal approximation. The lognormal approximation assumes that
lead time demand is distributed according to the lognormal distribution. This function
includes both the approaches outlined by Silver (1970) when P2<=0.9 and P2>0.9. These
approaches are available in popular textbooks e.g. Silver, Pyke and Thomas (2016, 311-312).
For P2<=0.9 see forthcoming paper by Saldanha (2022).

This function uses the input of a single lead time demand vector to calculate the ROP
for a single inventory item with a P2 proportion of demand fulfilled from available stock
fill rate under the continuous review (s, Q) inventory control policy using the lognormal
approximation. The lognormal approximation assumes that lead time demand is distributed
according to the lognormal distribution. This function includes both the approaches outlined
by Silver (1970) when P2<=0.9 and P2>0.9. Both of these functions are derived in the
forthcoming paper by Saldanha (2021).
}
\details{
The output is the single inventory items' lognormal ROP (s) if the roptru parameter is TRUE
or not entered (default) else if FALSE the SS for the P2 fill rate target is returned.
The demand should be a time series vector in the same units of time as lead time e.g.
if lead times are in weeks time series of demand should be aggregated weekly .

The output is the single inventory items' ROP or SS for the P2 fill rate target.
This function is primarily provided for use in theoretical studies and analysts must make
sure to ensure that the empirical data fits the lognormal distribution.
}
\examples{
CRlnorm.ld.fr(c(runif(24,2,8)),c(runif(24,20,40)),0.95,50)
This example returns a distributions with a mean of 100 and std. dev. of 120

CRlnorm.ld.fr(c(rlnorm(24,4.15917,0.944456)),0.95,50)
This example returns a distributions with a mean of 100 and std. dev. of 120

}
