% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRlnorm.x.fr.R
\name{CRlnorm.x.fr}
\alias{CRlnorm.x.fr}
\title{Lognormal Fill Rate ROP and SS Calculation from LTD data for the Continuous Review
Inventory Control Policy}
\usage{
CRlnorm.x.fr(x, p2, qty, roptru = TRUE)
}
\arguments{
\item{x}{vector of lead time demands}

\item{p2}{item's fill rate target proportion demand fulfilled from available stock}

\item{qty}{item's fixed order quantity}

\item{roptru}{if TRUE returns the ROP else the SS, default is the ROP}
}
\value{
the lognormal approximation reorder point (ROP)
}
\description{
This function uses the input of a single lead time demand vector to calculate the ROP
for a single inventory item with a P2 proportion of demand fulfilled from available stock
fill rate under the continuous review (s, Q) inventory control policy using the lognormal
approximation. The lognormal approximation assumes that lead time demand is distributed
according to the lognormal distribution. This function includes both the approaches outlined
by Silver (1970) when P2<=0.9 and P2>0.9. Both of these functions are derived in the
forthcoming paper by Saldanha (2021).
}
\details{
The output is the single inventory items' ROP or SS for the P2 fill rate target.
This function is primarily provided for use in theoretical studies and analysts must make
sure to ensure that the empirical data fits the lognormal distribution.
}
\examples{
CRlnorm.x.fr(c(rlnorm(24,4.15917,0.944456)),0.95,50)
# This example returns a distributions with a mean of 100 and std. dev. of 120

}
