% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRnorm.x.csl.R
\name{CRnorm.x.csl}
\alias{CRnorm.x.csl}
\title{Normal Cycle Service Level ROP and SS Calculation from lead time demand data under the
Continuous Review Inventory Control Policy}
\usage{
CRnorm.x.csl(x, p1, roptru = TRUE)
}
\arguments{
\item{x}{vector of lead time demands}

\item{p1}{item's cycle service level as p1-th PNS}

\item{roptru}{if TRUE returns the ROP else SS, default is TRUE}
}
\value{
The normal approximation ROP or SS
}
\description{
This function uses the input of a vector of lead time demand to calculate
the reorder point (ROP) or safety stock (SS) for a single inventory item with a P1
probability of no stockout (PNS) under the (s, Q) inventory control policy using the normal
approximation. The normal approximation assumes that lead time demand is distributed according
to the normal distribution. This function is available in popular textbooks e.g. Silver, Pyke
and Thomas (2016, 269-268).
}
\details{
The output is the single inventory items' normal ROP (s) if the roptru parameter is TRUE
or not entered (default) else if FALSE the SS for the target PNS P1 is returned.
}
\examples{
CRnorm.x.csl(c(runif(24,200,400)),0.95)
}
