% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRnorm.x.fr.R
\name{CRnorm.x.fr}
\alias{CRnorm.x.fr}
\title{Normal Fill Rate ROP or SS Calculation from LTD data for the Continuous Review Inventory
Control Policy}
\usage{
CRnorm.x.fr(x, p2, qty, roptru = TRUE)
}
\arguments{
\item{x}{vector of lead time demands}

\item{p2}{item's fill rate target proportion demand fulfilled from available stock}

\item{qty}{item's fixed order quantity}

\item{roptru}{if TRUE returns the ROP else SS, default is TRUE}
}
\value{
the normal approximation ROP or SS
}
\description{
This function uses the input of a single lead time demand vector to calculate the reorder
point (ROP) or safety stock (SS) for a single inventory item with a fill rate P2 proportion
of demand fulfilled from available stock under the continuous review (s, Q) inventory control
policy using the normal approximation. Note 's' is the ROP and Q is a fixed order quantity.
The normal approximation assumes that lead time demand is distributed according to the normal
distribution. This function includes both the approaches outlined by Silver (1970) when
P2<=0.9 and P2>0.9. These approaches are available in popular textbooks e.g.
Silver, Pyke and Thomas (2016, 271), equations 6.28 and 6.29.
}
\details{
The output is the single inventory items' normal ROP or SS for the P2 fill rate target
}
\examples{
CRnorm.x.fr(c(rnorm(24,100,120)),0.95,50)
}
