% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRunif.x.fr.R
\name{CRunif.x.fr}
\alias{CRunif.x.fr}
\title{Uniform Fill Rate ROP Calculation from LTD data for the Continuous Review Inventory
Control Policy}
\usage{
CRunif.x.fr(x, p2, qty, roptru = TRUE)
}
\arguments{
\item{x}{vector of lead time demands}

\item{p2}{item's fill rate target proportion demand fulfilled from available stock}

\item{qty}{item's fixed order quantity}

\item{roptru}{if TRUE returns the ROP else SS, default is TRUE}
}
\value{
The uniform approximation ROP or safety stock
}
\description{
This function uses the input of a single lead time demand vector to calculate the ROP
for a single inventory item with a P2 proportion of demand fulfilled from available stock
fill rate under the continuous review (s, Q) inventory control policy using the uniform
approximation. The uniform approximation assumes that lead time demand is distributed
according to the uniform distribution. This function includes both the approaches outlined
by Silver (1970) when P2<=0.9 and P2>0.9. Both of these functions are derived in the
forthcoming paper by Saldanha (2021).
}
\examples{
CRunif.x.fr(c(runif(24,66,134)),0.95,50)
}
