% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{demand_rates}
\alias{demand_rates}
\title{Compute demand rates from demand epochs and quantities}
\usage{
demand_rates(.data, date, qty)
}
\arguments{
\item{.data}{A tibble containing (at least) the following dates and quantities}

\item{date}{Epoch at which an order is placed}

\item{qty}{Quantity orded}
}
\value{
A tibble of demand rates by item
}
\description{
Compute demand rates from demand epochs and quantities
}
