% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densifiers.R
\name{densify}
\alias{densify}
\title{Calculate the nonparametric density of a numeric vector.}
\usage{
densify(x, type = c("ash", "kde"), trans = c("sqrt", "log", "none",
  "reflect"), ...)
}
\arguments{
\item{x}{numeric vector}

\item{type}{the type of density estimator (kde or ash)}

\item{trans}{transfrom to be applied to \code{x} before estimation (sqrt, log, none, reflect)}
}
\value{
a density as a tibble of \code{x}, \code{y} pairs
}
\description{
Calculate the nonparametric density of a numeric vector.
}
