% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densifiers.R
\name{density_to_pmf}
\alias{density_to_pmf}
\title{Make a pmf from a density}
\usage{
density_to_pmf(tbl)
}
\arguments{
\item{tbl}{a density object as a tibble of \code{x}, \code{y} pairs}
}
\value{
a pmf tibble of \code{x}, \code{prob}, and \code{cumprob}
}
\description{
Make a pmf from a density
}
