% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammfr.R
\name{gammfr}
\alias{gammfr}
\title{Gamma Fill Rate ROP and SS Calculation for the Continuous Review (s, Q) Inventory Control Policy}
\usage{
gammfr(qty, P2, muX, sigmX, Log)
}
\arguments{
\item{qty}{item's fixed order quantity}

\item{P2}{item's fill rate target proportion demand fulfilled from available stock}

\item{muX}{mean of lead time demand}

\item{sigmX}{standard devation of lead time demand}

\item{Log}{if TRUE returns the ROP else SS, default is TRUE}
}
\value{
The gamma approximation ROP or SS
}
\description{
This function calculates the reorder point (ROP) or safety stock (SS) for a single inventory
item with a P2 proportion of demand fulfilled from available stock fill rate under the (s, Q)
inventory control policy using the gamma approximation. The gamma approximation
assumes that lead time demand is distributed according to the gamma distribution. This function
includes both the approaches outlined by Silver (1970) when P2<=0.9 and P2>0.9.
For P2>0.9 the approach was derived by Tyworth, Guo, and Ganeshan (1996) and is also available
in popular textbooks e.g. Silver, Pyke and Thomas (2016, 311-312).
For P2<=0.9 see forthcoming paper by Foarth (2021).
}
\details{
The output is the single inventory items' gamma ROP (s) or SS for the P2 fill rate target
}
\examples{
gammfr(50,0.90,100,120)
}
