% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griding.R
\name{griding}
\alias{griding}
\title{Distribute emissions between areas and grids}
\usage{
griding(geoemiss, variable = NA, t_unit = NA, res = 5, type = "global",
  lat = c(-90, 90), lon = c(-180, 180), tol = res * 1e-05, verbose = T,
  plot = F, ...)
}
\arguments{
\item{geoemiss}{a output from set_total}

\item{variable}{name(s) of the pollutant(s)}

\item{t_unit}{time unit, defoult is "year"}

\item{res}{inventary resolution in degrees}

\item{type}{global, local or custon inventary type}

\item{lat}{latitude for custon inventory}

\item{lon}{lontitude for custon inventory}

\item{tol}{param passing to dist of sf::st_buffer}

\item{verbose}{display additional information}

\item{plot}{true for plot separate regions and final inventory}

\item{...}{aditional plot parameters}
}
\value{
a sfc with the grid geomatry and the emission by cell
}
\description{
calculate a inventory
}
\examples{
so2_no_2010 <- readRDS(paste0(system.file("extdata",package="inventory"),"/so2_no.Rds"))
# inventory for multiples areas
inventory_so2_2010 <- griding(so2_no_2010[1:3,], plot = TRUE)
# inventory for a single area and multiple pollutants
inventory_so2_2010 <- griding(so2_no_2010[1,], variable = c("so2","NO"), plot = TRUE)
# inventory for multiple areas and pollutants
inventory_so2_2010 <- griding(so2_no_2010[1:2,], variable = c("so2","NO"), plot = TRUE)

}
