% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalizers.R
\name{integrate_pdf}
\alias{integrate_pdf}
\title{Integrate (marginal) pdf to get cdf (as a function)}
\usage{
integrate_pdf(pdf)
}
\arguments{
\item{pdf}{marginal pdf to be integrated}
}
\description{
Integrate (marginal) pdf to get cdf (as a function)
}
