% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipe-tools.R
\name{inv_CheckRecipe}
\alias{inv_CheckRecipe}
\title{Recipe Checker}
\usage{
inv_CheckRecipe(.recipe_to_check, .inventory_to_check_against)
}
\arguments{
\item{.inventory_to_check_against}{}
}
\value{

}
\description{
This function takes a recipe specification and a reference
  inventory and returns the number of batches of that recipe which are
  possible. We can use this to determine which recipes in the book are
  possible with the current state of the inventory.

  Currently the specification simply checks that there is at least enough of
  every ingredient. When moving forward to potential repice planning or
  optimisation, it may become necessary to have the recipes expressed in a
  more complex data.type or as a function, to allow for multiple recipes to
  be solved at once, and to understand the interference with one another.
}
