% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{inv_ConnectInventory}
\alias{inv_ConnectInventory}
\title{Inventory connector}
\usage{
inv_ConnectInventory(
  input = NULL,
  .inventory_dir = get("inventory_dir"),
  filename = "inventory.yml"
)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
This connector either serves the value, if there is no input, or
  writes the input to the value of the inventory.
}
