% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{inv_ConnectRecipes}
\alias{inv_ConnectRecipes}
\title{Recipe Connector}
\usage{
inv_ConnectRecipes(
  input = NULL,
  .inventory_dir = get("inventory_dir"),
  filename = "recipe_book.yml"
)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
This connector either serves the value, if there is no input, or
  writes the input to the value of the recipe book.
}
