% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml-conversion.R
\name{inv_ConvertFromYAML}
\alias{inv_ConvertFromYAML}
\title{Convert From YAML}
\usage{
inv_ConvertFromYAML(yaml_list_in)
}
\arguments{
\item{yaml_list_in}{}
}
\value{

}
\description{
This function captures the key transformation of the inventory
  list from the YAML storage specification into the unit rich R specification
  which is used internally.
}
