% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yaml-conversion.R
\name{inv_ConvertToYAML}
\alias{inv_ConvertToYAML}
\title{Convert To YAML}
\usage{
inv_ConvertToYAML(input_list)
}
\arguments{
\item{input_list}{}
}
\value{

}
\description{
This function takes the internal storage format, including
  units, and reformats them so that they can be encoded in the YAML storage
  format used by the package at present.

  While in future alternative storage formats may be included, for the
  limited single user use cases which are currently present, YAML is
  currently sufficient, and will remain with the package for similar use
  cases in future.
}
