% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_printer.R
\name{inv_PrintInventory}
\alias{inv_PrintInventory}
\title{Inventory Printer Function}
\usage{
inv_PrintInventory(.inventory = inv_ConnectInventory())
}
\arguments{
\item{.inventory}{}
}
\value{

}
\description{
Function to print a copy of the inventory in a nice tibble
  format.
}
