% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipe-tools.R
\name{inv_SelectRecipe}
\alias{inv_SelectRecipe}
\title{Recipe Selection}
\usage{
inv_SelectRecipe(.recipe_book = test_recipes, .inventory = test_inventory)
}
\arguments{
\item{.inventory}{}
}
\value{

}
\description{
This function takes the recipe book and the inventory provided
  and returns the name of a recipe in the recipe book which will be possible
  when considering the ingredients required and the stock in the provided
  inventory.
}
