% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipe-tools.R
\name{inv_Unitify}
\alias{inv_Unitify}
\title{Fundamental Specification to Unit Transform}
\usage{
inv_Unitify(list_to_unit)
}
\arguments{
\item{list_to_unit}{}
}
\value{

}
\description{
This function encodes the transformation from the conventional
  storage model used in this project into a unit rich value which can be used
  in the internal calculations of the inventoRy package.
}
