% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R, R/inventory_package.R
\docType{package}
\name{inventory}
\alias{inventory}
\alias{inventory-package}
\title{inventory function}
\usage{
inventory(forecast, h, ipolicy = "MTS")
}
\arguments{
\item{forecast}{A data frame or matrix which represents the forecast
distribution}

\item{h}{The forecast horizon}

\item{ipolicy}{The inventory policy. "MTS" (Make to Stock) by default}
}
\value{
A list of average on-hand inventory and service levels.
}
\description{
Creates simulation from a given forecast

This package provides functionality for a inventory simulations via a Monte
Carlo setup for robust forecast evaluation.
}
\details{
This function creates a inventory simulation based on a forecast and a
given inventory policy to evaluate the average on-hand inventory and the
service level.
}
\examples{
\dontrun{
   inventory(forecastdata, h=12, ipolicy="MTS")
}
}
\author{
Yves R. Sagaert

import forecast

Yves R. Sagaert \email{yves.r.sagaert@gmail.com}
}
