% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_curve.R
\name{inventory_curve}
\alias{inventory_curve}
\title{Inventory Curve Plot}
\usage{
inventory_curve(
  models_inv,
  modelnames,
  colx = "scaled_av_onhand_inventory",
  coly = "fill_rate",
  xlab = "Average on hand inventory",
  ylab = "1 - Fill Rate",
  zoom = TRUE,
  line_type = c(1, 2, 3),
  line_width = rep(1, length(modelnames)),
  line_pch = c(0, 1, 2, 5, 6, 4, 3),
  line_colors = c("#7FC97F", "#BEAED4", "#FDC086", "#FFFF99", "#386CB0", "#F0027F",
    "#BF5B17")
)
}
\arguments{
\item{models_inv}{A list of matrices with the results from inventory
evaluation.}

\item{modelnames}{A vector with the names of the model for the legend of
the plot.}

\item{colx}{Data used on the x-axis. Default set on
"scaled_av_onhand_inventory". This can also be set on "av_onhand_inventory"
or "av_inventory".}

\item{coly}{Data used on the y-axis. Default set on "fill_rate", which will
then show 1-Fill Rate. Other options are "av_shortage_items" and
"scaled_av_shortage_items".}

\item{xlab}{Label used on the x-axis. Default set on "Average on hand
inventory".}

\item{ylab}{Label used on the y-axis. Default set on "1 - Fill Rate".}

\item{zoom}{Boolean variable to set the plot to zoom to the relevant data.
Default is set on TRUE.}

\item{line_type}{Type of lines used in the plot.}

\item{line_width}{Width of the lines in the plot. Default set on 1.}

\item{line_pch}{A vector with the symbols used for every line.
Default vector provided.}

\item{line_colors}{A vector with colors used for the lines in the plot.
Default vector provided.}
}
\value{
A plot with the trade-off curves between service level (fill rate)
and inventory (average onhand inventory).
}
\description{
Plot a trade-off curve between fill rate and average on-hand inventory.
}
\details{
The plot shown visualised a curve for every model via a sequence of alphas
where the 1-fill rate is shown on the y-axis and the average on-hand
inventory is shown at the x-axis.
}
\examples{
\dontrun{
set.seed(101010)
series <- rnorm(60,100,1)
test_series <- rnorm(12,100,10)
forecast <- series[(length(series)-12):length(series)]
ri1 <- alphas_inventory_evaluation(demand_forecast=list(forecast,sd=rep(3,12)),
demand_test=test_series, hist_demand=series, plot.out=TRUE)
ri2 <- alphas_inventory_evaluation(demand_forecast=list(forecast-5,sd=rnorm(12,6,2)),
demand_test=test_series, hist_demand=series, plot.out=TRUE)
inventory_curve(models_inv=list(ri1,ri2),
modelnames=c("Trial model1","Trial model2"),zoom=FALSE)
}
}
\author{
Yves R. Sagaert
}
