% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{lead_times}
\alias{lead_times}
\title{Calculate lead times from work-order dates and completed dates}
\usage{
lead_times(.data, start, end)
}
\arguments{
\item{.data}{A tibble containing (at least) dates for work order placement and completion}

\item{start}{Date, start of lead time}

\item{end}{Date, end of lead time}

\item{item}{Item id}
}
\value{
.data with lead-time column appended
}
\description{
Calculate lead times from work-order dates and completed dates
}
