% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormfr.R
\name{lnormfr}
\alias{lnormfr}
\title{Lognormal Fill Rate ROP Calculation for the Continuous Review Inventory Control Policy}
\usage{
lnormfr(qty, P2, muX, sigmX)
}
\arguments{
\item{qty}{item's fixed order quantity}

\item{P2}{item's fill rate target proportion demand fulfilled from available stock}

\item{muX}{first moment (mean) of lognormally distributed lead time demand}

\item{sigmX}{second moment (standard devation) of lognormally distributed lead time demand}
}
\value{
the lognormal approximation reorder point (ROP)
}
\description{
This function calculates the ROP for a single inventory item with a P2 proportion
of demand fulfilled from available stock fill rate under the continuous review (s, Q)
inventory control policy using the lognormal approximation. The lognormal approximation assumes
that lead time demand is distributed according to the lognormal distribution. This function
includes both the approaches outlined by Silver (1970) when P2<=0.9 and P2>0.9.
Both of these functions are derived in the forthcoming paper by Foarth (2021).
}
\details{
This function is primarily provided for use in theoretical studies and analysts must make sure
to convert the standard means (mean) and standard deviations (sd) to the first and second moments
of the lognormal distribution.
}
\examples{
lnormfr(50,0.95,4.585559829,0.1980422)
This example returns a distributions with an average of 100 and corresponding std. dev. of 120
}
