% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_series_inventory_evaluation.R
\name{multiple_series_inventory_evaluation}
\alias{multiple_series_inventory_evaluation}
\title{Inventory Evaluation for multiple series}
\usage{
multiple_series_inventory_evaluation(
  model_forecasts,
  demand_test,
  alpha = 0.05,
  plot.out = FALSE,
  mode = "normal",
  hist_demand = NA,
  alpha.seq = TRUE
)
}
\arguments{
\item{model_forecasts}{A list of different model forecasts. An individual
model forecast can be either a forecast object, a list consisting
of a mean and sd vector for the forecast horizon, or a point cloud of the
distribution of each forecast horizon.}

\item{demand_test}{A vector containing the actual demand for the inventory
evaluation. The length of the test demand should be equal to the length of
the demand forecast.}

\item{alpha}{A sequence of target service levels of the inventory simulation.
This is can be a single value, or a sequence. If alpha.seq is set on TRUE, a
sequence is generated (by default).}

\item{plot.out}{Parameter to plot a graphic for every alpha in the sequence
, default set on FALSE.}

\item{mode}{Shape of the distribution, default set on 'normal'.}

\item{hist_demand}{A vector of historical demand of the series, where the demand
forecast model is build on. Default is set on NA.}

\item{alpha.seq}{Should the function generate a default alpha sequence. This
is set by default on TRUE.}
}
\value{
A numeric vector with: used service level parameter alpha, stockout
rate, average inventory, average on-hand inventory, fill-rate, average test
demand, ratio average on-hand inventory on average test demand. A plot of
the simulation can be shown.
}
\description{
Evaluates a rolling forecast via inventory evaluation for multiple series.
}
\details{
This function evaluates rolling forecasts for multiple series for a sequence
of alphas via an inventory simulation. The output can be used to generate
a trade-off curve between fill rate and average on-hand inventory for one
specific series or averaged over all series.
}
\examples{
\dontrun{
series <- rnorm(60,100,1)
test_series <- rnorm(12,100,5)
forecast1 <- series[(length(series)-12):length(series)]
forecast2 <- rep(mean(series),12)
modelfc1 <- list(forecast1,sd=rep(3,12))
modelfc2 <- list(forecast2,sd=rep(1,12))
inventory(model_forecasts=list(modelfc1, modelfc2),
demand_test=test_series, hist_demand=series, plot.out=FALSE)
}

}
\author{
Yves R. Sagaert
}
