% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normfr.R
\name{normfr}
\alias{normfr}
\title{Normal Fill Rate ROP or SS Calculation for the Continuous Review Inventory Control Policy}
\usage{
normfr(qty, P2, muX, sigmX, Log = TRUE)
}
\arguments{
\item{qty}{item's fixed order quantity}

\item{P2}{item's fill rate target proportion demand fulfilled from available stock}

\item{muX}{mean of lead time demand}

\item{sigmX}{standard devation of lead time demand}

\item{Log}{if TRUE returns the ROP else SS, default is TRUE}
}
\value{
the normal approximation ROP or SS
}
\description{
This function calculates the reorder point (ROP) or safety stock (SS) for a single inventory
item with a P2 proportion of demand fulfilled from available stock fill rate under the
continuous review (s, Q) inventory control policy using the normal approximation. Note 's' is
the ROP and 'Q' is a fixed order quantity. The normal approximation assumes that lead time
demand is distributed according to the normal distribution. This function includes both the
approaches outlined by Silver (1970) when P2<=0.9 and P2>0.9. These approaches are available
in popular textbooks e.g. Silver, Pyke and Thomas (2016, 271), see equations 6.28 and 6.29.
}
\details{
The output is the single inventory items' normal ROP or SS for the P2 fill rate target
}
\examples{
normfr(50,0.95,100,120)
}
