% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{orders_to_rates}
\alias{orders_to_rates}
\title{Calculate the nonparametric density of a numeric vector.}
\usage{
orders_to_rates(.data, date, qty, ...)
}
\arguments{
\item{.data}{tibble containing dates and quantities}

\item{date}{order-date column}

\item{qty}{order-quantity column}

\item{...}{grouping variable(s) (eg item)}
}
\value{
a demand-rate transformed tibble of (eg, demand) rates
}
\description{
Calculate the nonparametric density of a numeric vector.
}
