% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{outp}
\alias{outp}
\title{Order-Up-To level inventory simulation function}
\usage{
outp(data, fcs, holdout, ss = c("constant", "dynamic"), L = 1, CSL = 95)
}
\arguments{
\item{data}{Demand observations, both insample and holdout (simulation) windows.}

\item{fcs}{Lead time forecasts (aggregated), equal size as demand data.}

\item{holdout}{The size of the window where the inventory simulation will be performed.}

\item{ss}{Safety stock; single value, vector, or an option between "constant" or "dynamic".
If single value, then ss is considered fixed equal to that value. If vector, then ss is
considered to be updated dynamically; in this case, the length of ss should be equal to
the holdout. If "constant", then ss is calculated empirically once (based on the insample
data) and kept fixed. If "dynamic", then ss is calculated empirically and updated for
each period. It is recommended that both "constant" and "dynamic" are used only if at
least 20 observations are available in the insample (length of data minus holdout).}

\item{L}{Lead time.}

\item{CSL}{Targeted Customer Service Level.}
}
\value{
The list, containing the following, is returned:
\itemize{
\item Orders variance
\item Inventory variance
\item Total (inventory and backlog) cost
\item Realised CSL
}
}
\description{
This is the first inventory simulation function.
}
\details{
Function uses data and rolling origin forecasts in order to do simulations
}
\examples{
#No example here yet.

}
