% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possess.R
\name{possess}
\alias{possess}
\title{Read a georeference image and extract values for using in griding}
\usage{
possess(geoemiss, filename = NA, plots = T, verbose = T)
}
\arguments{
\item{geoemiss}{an output from geoemiss}

\item{filename}{image filename}

\item{plots}{plot individual regions}

\item{verbose}{display additional information}
}
\description{
extract values from a image to be used on each region
}
\examples{
states <- sf::st_read(paste(system.file("extdata", package = "inventory"),"/states.shp",sep=""),
                      quiet = TRUE)
Nox    <- geoemiss(geom = states,variable = "Nox",names = c("sp","rj"),values = c(1000,20))
image  <- paste(system.file("extdata", package = "inventory"),"/tiny.tif",sep="")
ras    <- possess(Nox,image,plots = FALSE)
test   <- griding(ras,variable = "Nox", res = 0.1, type = "local", plot = TRUE)
}
