% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densifiers.R
\name{rescale_pdf}
\alias{rescale_pdf}
\title{Rescale pdf, f_X to f_Y where y = rx,}
\usage{
rescale_pdf(rate, pdf)
}
\arguments{
\item{rate}{the factor by which to rescale the pdf}

\item{pdf}{the pdf to be rescaled, as a _function_}
}
\value{
a rescaled pdf as a _function_
}
\description{
Rescale pdf, f_X to f_Y where y = rx,
}
