% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin_inventory.R
\name{rolling_origin_inventory}
\alias{rolling_origin_inventory}
\title{Rolling Origin experiment with Inventory Evaluation}
\usage{
rolling_origin_inventory(
  data = "M3",
  models = c("Naive", "ETS"),
  plot.ro = TRUE
)
}
\arguments{
\item{data}{Example dataset that will be used to perform rolling origin
experiment. Default set on "M3".}

\item{models}{A vector with all the different model names. Default is set
to c("Naive", "ETS").}

\item{plot.ro}{Parameter to plot a forecast plot of every model with the
rolling origin forecast and prediction intervals.}
}
\value{
A matrix with the inventory evaluation results for all series, for
all models, for a predefined sequence of alphas. The results are summarised
by stockout ratio, average inventory, average onhand inventory, fill rate,
average test demand, scaled onhand inventory, average shortage items and
scaled average shortage items. The scaling happens with the average test
demand.
}
\description{
Make a rolling origin setup with a recurring forecast for different models
and evaluate these forecasts on inventory performance.
}
\details{
This function makes the setup of a rolling origin for an example dataset.
For different models, a forecast is made for every rolling origin. These
forecasts can be plot, but are then evaluated on their inventory performance
via an inventory simulation.
}
\examples{
\dontrun{
rolling_origin_inventory(data="M3", model=c("Naive","ETS"), plot.ro=TRUE)
}

}
\author{
Yves R. Sagaert
}
