% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalizers.R
\name{rop_fun}
\alias{rop_fun}
\title{Return reorder-point function by inverting cdf}
\usage{
rop_fun(cdf)
}
\arguments{
\item{cdf}{Marginal lead-time demand cdf, as a _function_}
}
\value{
A function that returns a two-column tibble with
    vars: \code{svc_lvl} and \code{rop}
}
\description{
Return reorder-point function by inverting cdf
}
