% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densifiers.R
\name{safely_densify}
\alias{safely_densify}
\title{Calculate density (safely) of a numeric vector.}
\usage{
safely_densify(...)
}
\arguments{
\item{x}{numeric vector}

\item{type}{the type of density estimator (kde or ash)}

\item{trans}{transfrom to be applied to \code{x} before estimation (sqrt, log, none, reflect)}
}
\value{
a density as a tibble of \code{x}, \code{y} pairs
}
\description{
Calculate density (safely) of a numeric vector.
}
