% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_inventory.R
\name{save_inventory}
\alias{save_inventory}
\title{Save the inventory in NetCDF format}
\usage{
save_inventory(gi, filename = NA, dates, variable, unit = NA, mw = 1,
  COMPRESS = NA, force_ncdf4 = F, verbose = T)
}
\arguments{
\item{gi}{a output from inventory of battery (multiple times)}

\item{filename}{name of the file}

\item{dates}{date(s) for the data}

\item{variable}{name(s) of the pollutant(s) to write}

\item{unit}{inventory unit}

\item{mw}{molecular weight}

\item{COMPRESS}{compression level from 1 to 9, or NA for no compression}

\item{force_ncdf4}{force NetCDF4 format}

\item{verbose}{display additional information}
}
\description{
Create a NetCDF file from a 'inventoty' output and can add new variables to
}
\examples{
grinded_so2 <- sf::read_sf(paste0(system.file("extdata",package="inventory"),"/grid_so2.gpkg"))
dir.create(file.path(tempdir(), "INV"))
save_inventory(grinded_so2,filename = paste0(file.path(tempdir(), "INV"),"test.nc"),
                variable = "so2", dates = '2010-01-01')
}
