% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_inventory_distance.R
\name{scaled_inventory_distance}
\alias{scaled_inventory_distance}
\title{Minimum Scaled Inventory Distance}
\usage{
scaled_inventory_distance(inventory_result, scaled = TRUE)
}
\arguments{
\item{inventory_result}{A matrix with the results from inventory evaluation.}

\item{scaled}{Use the scaled SID by default. The use of unscaled inventory
distance may skew results to series with large shortages or inventory.
Default set on TRUE.}
}
\value{
Numeric value of SID.
}
\description{
Calculate the inventory evaluation metric SID
}
\details{
SID is the minimum scaled inventory distance which makes the trade-off between
fill-rate service level and average onhand inventory.
}
\examples{
example_inventory_result=data.frame(m=c(1,1,1,2,2,2),
av_onhand_inventory=c(10,11,12,20,25,30),
fill_rate=c(0.9,0.92,0.95,1,1,1),
scaled_av_onhand_inventory = c(0.20, 0.22, 0.24, 0.40, 0.50, 0.60),
av_shortage_items = c(5,4,2,0,0,0),
scaled_av_shortage_items=c(0.10, 0.08, 0.04, 0,0,0))

scaled_inventory_distance(inventory_result=example_inventory_result)

scaled_inventory_distance(inventory_result=example_inventory_result, scaled=FALSE)

}
\author{
Yves R. Sagaert
}
