% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacker.R
\name{stacker}
\alias{stacker}
\title{Agrup sequencial emissions for record}
\usage{
stacker(grid, variable = NA, verbose = TRUE)
}
\arguments{
\item{grid}{list of ouputs from griding funcion}

\item{variable}{name(s) of the pollutant(s)}

\item{verbose}{display additional information}
}
\description{
create stack of grinded emissions for record in a NetCDF file with multiple times (dates)
}
\examples{
grinded_so2 <- readRDS(paste0(system.file("extdata",package="inventory"),"/grid_so2.Rds"))
two_times_grinded_so2 <- stacker(list(a = grinded_so2, b = grinded_so2))
}
