% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tableize_demand}
\alias{tableize_demand}
\title{Compute daily demand distribution}
\usage{
tableize_demand(.data, date, qty)
}
\arguments{
\item{.data}{A tibble containing (at least) the following dates and quantities}

\item{date}{Quoted column name within `.data` referring to the Date on which demand is to be met (and hence, inventory consumed)}

\item{qty}{Quoted column name within `.data` referring to the quantity orded (_i.e._, _demanded_)}
}
\value{
A frequency tibble of daily demand
}
\description{
Compute daily demand distribution
}
