% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tableize_demand2}
\alias{tableize_demand2}
\title{Compute daily demand distribution}
\usage{
tableize_demand2(.data, date_range, order_date, qty)
}
\arguments{
\item{.data}{A tibble containing (at least) the following dates and quantities}

\item{date_range}{Overall date range spanned by the union of all `Item Numbers` in the data set`}

\item{qty}{Quoted column name within `.data` referring to the quantity orded (_i.e._, _demanded_)}

\item{date}{Quoted column name within `.data` referring to the Date on which demand is to be met (and hence, inventory consumed)}
}
\value{
A frequency tibble of daily demand
}
\description{
Compute daily demand distribution
}
