% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{test_history}
\alias{test_history}
\title{A testing Dataset for the history}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.
}
\usage{
test_history
}
\description{
A testing Dataset for the history
}
\keyword{datasets}
