% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{test_inventory}
\alias{test_inventory}
\title{A testing Dataset for the inventory}
\format{
An object of class \code{list} of length 1.
}
\usage{
test_inventory
}
\description{
A testing Dataset for the inventory
}
\keyword{datasets}
