% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{test_recipes}
\alias{test_recipes}
\title{A testing Dataset for the recipes}
\format{
An object of class \code{list} of length 1.
}
\usage{
test_recipes
}
\description{
A testing Dataset for the recipes
}
\keyword{datasets}
