% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnormfr.R
\name{tnormfr}
\alias{tnormfr}
\title{Truncated Normal Fill Rate ROP Calculation for the Continuous Review Inventory Control Policy}
\usage{
tnormfr(qty, P2, muX, sigmX)
}
\arguments{
\item{qty}{item's fixed order quantity}

\item{P2}{item's fill rate target proportion demand fulfilled from available stock}

\item{muX}{mean of lead time demand}

\item{sigmX}{standard devation of lead time demand}
}
\value{
the truncated normal approximation reorder point (ROP)
}
\description{
This function calculates the ROP for a single inventory item with a P2 proportion
of demand fulfilled from available stock fill rate under the continuous review (s, Q)
inventory control policy using the truncated normal approximation. The truncated normal
approximation assumer that lead time demand is distributed according to the truncated normal
 distribution. This function includes both the approaches outlined by Silver (1970)
when P2<=0.9 and P2>0.9. Both of these functions are derived in the forthcoming paper
by Foarth (2021).
}
\examples{
tnormfr(50,0.95,100,120)
}
