% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniffr.R
\name{uniffr}
\alias{uniffr}
\title{Uniform Fill Rate ROP Calculation for the Continuous Review Inventory Control Policy}
\usage{
uniffr(qty, P2, lb0, ub0)
}
\arguments{
\item{qty}{item's fixed order quantity}

\item{P2}{item's fill rate target proportion demand fulfilled from available stock}

\item{lb0}{the lower bound of the uniform distribution}

\item{ub0}{the upper bound of the uniform distribution}
}
\value{
The uniform approximation reorder point
}
\description{
This function calculates the ROP for a single inventory item with a P2 proportion
of demand fulfilled from available stock fill rate under the continuous review (s, Q)
inventory control policy using the unform approximation. The uniform approximation assumes
that lead time demand is distributed according to the uniform distribution. This function
includes both the approaches outlined by Silver (1970) when P2<=0.9 and P2>0.9.
Both of these functions are derived in the forthcoming paper by Foarth (2021).
}
\examples{
uniffr(50,0.95,66,136)
}
