% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_inventory.R
\name{write_inventory}
\alias{write_inventory}
\title{Save the inventory in NetCDF format}
\usage{
write_inventory(gi, filename = NA, dates, variable, unit = NA, mw = 1,
  COMPRESS = NA, force_ncdf4 = F, verbose = T)
}
\arguments{
\item{gi}{a output from inventory}

\item{filename}{name of the file}

\item{dates}{date for the data}

\item{variable}{name(s) of the pollutant(s) to write}

\item{unit}{inventory unit}

\item{mw}{molecular weight}

\item{COMPRESS}{compression level from 1 to 9, or NA for no compression}

\item{force_ncdf4}{force NetCDF4 format}

\item{verbose}{display additional information}
}
\description{
Create a NetCDF file from a 'inventoty' output
}
\examples{
grinded_so2 <- sf::read_sf("C:/Users/Schuch/Desktop/test_inventarios/grid_so2_new.gpkg")
dir.create(file.path(tempdir(), "EMISS"))
write_inventory(grinded_so2,filename = paste0(file.path(tempdir(), "EMISS"),"test.nc"),
                variable = "so2", dates = '2010-01-01')
}
