% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRboot.ld.csl.R
\name{CRboot.ld.csl}
\alias{CRboot.ld.csl}
\title{Estimate Bootstrap ROP or SS from Lead-Time and Demand Data for a Continuous Review Policy
with a Target Cycle Service Level}
\usage{
CRboot.ld.csl(x, y, p1, B = 500, roptru = TRUE, seed = as.numeric(Sys.time()))
}
\arguments{
\item{x}{a vector of lead-times}

\item{y}{a vector of demands}

\item{p1}{the cycle service level as PNS}

\item{B}{the number of bootstrap resamples, default is 500}

\item{roptru}{if TRUE returns the ROP else the SS, default is the ROP}

\item{seed}{a random seed, default is the computer time}
}
\value{

}
\description{

}
\details{
The ROP is the default value returned from the function but if safety stock is desired
then an optional input can be added using "roptru=FALSE" will return the safety stock.

Note that the bootstrap approach utilizes a random resampling of the input lead time and
demand resamples with replacement. The random seed for the random resampling is set to the
current machine's system time. Hence, if replicability of the results are desired e.g., the
ROP and safety stock for a number of items need to be calculated then it is recommended to
use the input "seed=" (see examples).
}
\examples{
a<-rnorm(24,5,1)
b<-rnorm(24,20,5)
Returns the ROP using 500 bootstrap resamples with a PNS of 95\%
CRboot.ld.csl(a,b,.95,500)
Returns the SS using 500 bootstrap resamples with a PNS of 95\%
CRboot.ld.csl(a,b,50,.95,500,FALSE)
Returns the ROP and SS from the same random numbers
CRboot.ld.csl(a,b,50,.95,500,TRUE,0)
CRboot.ld.csl(a,b,50,.95,500,FALSE,0)
}
