% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{generate_ddlt}
\alias{generate_ddlt}
\title{Generate demand-during-lead-time distribution}
\usage{
generate_ddlt(lead_times, demand_freq, cycles = 999)
}
\arguments{
\item{lead_times}{A numeric vector of lead times in days}

\item{demand_freq}{A frequency tibble of daily demands (normally gotten as output from tableize_demand()`)}

\item{cycles}{Number of 'saw tooth' inventory replenishment cycles to simulate}
}
\value{
A numeric vector of demands during lead time, of length `cycles`
}
\description{
Generate demand-during-lead-time distribution
}
