% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoemiss.R
\name{geoemiss}
\alias{geoemiss}
\title{Gather data for calculate the inventory}
\usage{
geoemiss(geom, values, variable, names = NA, mass_unit = "t", verbose = T)
}
\arguments{
\item{geom}{sf object describing the geometry of each area}

\item{values}{numeric vector (for one variable) or list of vectors (for multiple variables) of total mass}

\item{variable}{name(s) of the pollutant(s)}

\item{names}{area names}

\item{mass_unit}{mass unit, default is "t" (ton)}

\item{verbose}{display additional information}
}
\value{
a sf with the area shapes and the total emission for each area
}
\description{
create a sf object with the emission and the geographic area which this emission
will be distributed.
}
\examples{
continents <- sf::st_read(paste0(system.file("extdata",package="inventory"),"/continent.shp"),
                          quiet = TRUE)
continents <- continents[1:5,]
so2 <- read.csv(paste0(system.file("extdata",package="inventory"),"/SO2.csv"))
names(so2) <- c("region","year","mass")
so2 <- so2[so2$year == 2010,]
# for a single pollutant
so2_2010 <- geoemiss(geom     = continents,
                     values   = so2$mass,
                     names    = so2$region,
                     variable = "so2")
# for multiples pollutants, NO values is twice so2 emissions for exercice purpose
so2_NO_2010 <- geoemiss(geom     = continents,
                        values   = list(so2$mass,2 * so2$mass),
                        names    = so2$region,
                        variable = c("so2","NO"))
}
