% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\docType{package}
\name{inventory}
\alias{inventory}
\alias{inventory-package}
\title{A set of tools used to compute optimal reorder points.}
\description{
A set of tools used to compute optimal reorder points.
}
