% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory}
\alias{inventory}
\title{Distribute emissions between areas and grids}
\usage{
inventory(totals, variable = NA, t_unit = NA, verbose = T, plot = F,
  ...)
}
\arguments{
\item{totals}{a output from set_total}

\item{variable}{name(s) of the pollutant(s)}

\item{t_unit}{time unit, defoult is "year"}

\item{verbose}{display additional information}

\item{plot}{true for plot separate regions and final inventory}

\item{...}{aditional plot parameters}
}
\value{
a sfc with the grid geomatry and the emission by cell
}
\description{
calculate a inventory
}
\examples{
# multiples areas
so2_2010   <- readRDS("C:/Users/Schuch/Desktop/test_inventarios/totals_01.Rds")
inventory_so2_2010 <- inventory(so2_2010[1:2,], plot = TRUE)
# multiple pollutants
so2_no_2010   <- readRDS("C:/Users/Schuch/Desktop/test_inventarios/so2_no.Rds")
inventory_so2_2010 <- inventory(so2_no_2010[1,], variable = c("so2","NO"), plot = TRUE)
# multiple areas and pollutants
so2_no_2010   <- readRDS("C:/Users/Schuch/Desktop/test_inventarios/so2_no.Rds")
inventory_so2_2010 <- inventory(so2_no_2010[1:3,], variable = c("so2","NO"), plot = TRUE)

}
