% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R, R/inventory_package.R
\docType{package}
\name{inventory}
\alias{inventory}
\alias{inventory-package}
\title{Inventory}
\usage{
inventory(forecast)
}
\arguments{
\item{forecast}{A forecast object}
}
\value{
A data frame
}
\description{
Evaluates forecast on inventory performance

This packages provides functionalities to evaluate a demand forecast on
inventory performance via inventory simulation. A trade-off is shown between
the service level / fill rate and the average on-hand inventory.
}
\details{
This function returns the output of an inventory simulation based on the
forecast input.
}
\examples{
series <- rnorm(60,100,1)
inventory(series)

}
\author{
Yves R. Sagaert

Yves R. Sagaert \email{yves.r.sagaert@gmail.com}
}
