% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Read and write metadata}
\usage{
metadata(filename = NA, variable = 0, attname = NA, action = "read",
  value = NA, verbose = F)
}
\arguments{
\item{filename}{file name}

\item{variable}{variable name, 0 to global and "?" to ask all names}

\item{attname}{attribute names}

\item{action}{Read or write attribute (NA for get all attnames)}

\item{value}{value to write}

\item{verbose}{display additional information}
}
\description{
Read and write metadata information of a NetCDF inventoty file
}
\examples{

}
