% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possess.R
\name{possess}
\alias{possess}
\title{Read a georeference image and extract values for using in griding}
\usage{
possess(geoemiss, filename = NA, plots = T, verbose = T)
}
\arguments{
\item{geoemiss}{a output from geoemiss}

\item{filename}{image filename}

\item{plots}{plot individual regions}

\item{verbose}{display additional information}
}
\description{
extract values from a image to be used on each region
}
\examples{
so2_no_2010 <- readRDS(paste0(system.file("extdata",package="inventory"),"/so2_no.Rds"))
file <- paste(system.file("extdata", package = "EmissV"),"/dmsp.tiff",sep="")
# a_raster <- possess(so2_no_2010,file)

}
