% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveInventory.R
\name{saveInventory}
\alias{saveInventory}
\title{Save the inventory in NetCDF format}
\usage{
saveInventory(gi, filename = NA, dates, variable, unit = NA, mw = 1,
  time_unit = "year", COMPRESS = NA, force_ncdf4 = F, verbose = T)
}
\arguments{
\item{gi}{a output from inventory of battery (multiple times)}

\item{filename}{name of the file}

\item{dates}{date(s) for the data}

\item{variable}{name(s) of the pollutant(s) to write}

\item{unit}{inventory unit}

\item{mw}{molecular weight}

\item{time_unit}{time unit, defoult is "year"}

\item{COMPRESS}{compression level from 1 to 9, or NA for no compression}

\item{force_ncdf4}{force NetCDF4 format}

\item{verbose}{display additional information}
}
\description{
Create a NetCDF file from a 'inventoty' output and can add new variables to
}
\examples{
grinded_so2 <- readRDS(paste0(system.file("extdata",package="inventory"),"/grid_so2.Rds"))
dir.create(file.path(tempdir(), "INV"))
saveInventory(grinded_so2,filename = paste0(file.path(tempdir(), "INV"),"test.nc"),
                variable = "so2", dates = '2010-01-01')
}
