% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_total.R
\name{set_total}
\alias{set_total}
\title{prepate data for inventoryes}
\usage{
set_total(geom, variable, var_name = NA, names = NA, unit = NA,
  verbose = T)
}
\arguments{
\item{geom}{geo}

\item{variable}{list}

\item{names}{names}

\item{unit}{unit, defoult is ton by year "t/year"}

\item{verbose}{display additional information}
}
\value{
a sfc with the names, total of pollutants and geometry to make the inventory
}
\description{
Organize inventory data
}
\examples{
continents <- sf::read_sf(paste0(system.file("extdata",package="inventory"),"/continent.shp"))
continents <- continents[1:5,]
so2 <- read.csv(paste0(system.file("extdata",package="inventory"),"/SO2.csv"))
names(so2) <- c("region","year","mass")
so2 <- so2[so2$year == 2010,]
totals <- set_total(geom     = continents,
                    variable = so2$mass,
                    names    = so2$region,
                    var_name = "so2")
}
