% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacker.R
\name{stacker}
\alias{stacker}
\title{Agroup sequencial emissions for record}
\usage{
stacker(grid, variable = NA, verbose = TRUE)
}
\arguments{
\item{grid}{list of sf from griding function}

\item{variable}{name(s) of the pollutant(s)}

\item{verbose}{display additional information}
}
\value{
a sf with the grid geometry and the emission by cell for multiple times
}
\description{
Create a stack of multiple grinded emissions for record in a NetCDF file with multiple times (dates)
}
\examples{
grinded_so2 <- readRDS(paste0(system.file("extdata",package="inventory"),"/grid_so2.Rds"))
two_times_grinded_so2 <- stacker(list(a = grinded_so2, b = grinded_so2))
}
